SudokuSolver
Added the SudokuSolver class. This class contains the main() method
which reads the user input from file specified by the user. An absolute
file path can be provided otherwise the file is looked for in the “src”
folder in which the SudokuSolver is located. The string read from the
input file is translated into the Initial State and the solver is
called with the initial state. The A* search is timed.
The search time and the final solution are printed on the console.